Refactor majority of the code
