+++ src/content/vars.js
@@ -1,0 +1,4 @@
+let variables = {};
+let messages = {};
+let prefix = 'mji-';
+let postponeAction = 'none';
