+++ src/options/options.html
@@ -4,51 +4,7 @@
         <meta charset="utf-8">
         <script type="module" src="/lib/browser-polyfill.min.js"></script>
         <script type="module" src="./options.js"></script>
-        <style>
-            .panel {
-                padding: 2em;
-            }
-
-            .subpanel {
-                margin-top: 1em;
-                margin-left: 3em;
-                margin-bottom: 2em;
-            }
-
-            .group:not(:last-child) {
-                margin-bottom: 1em;
-            }
-
-            textarea, .textinput {
-                width: 100%;
-            }
-
-            #cached-messages {
-                margin-top: 0.5em;
-            }
-
-            textarea {
-                resize: vertical;
-                font-family: monospace;
-                height: 500px;
-            }
-
-            label {
-                vertical-align: middle;
-            }
-
-            .hidden {
-                display: none;
-            }
-
-            .important {
-                color: #aa2222;
-            }
-
-            .spacer {
-                height: 1em;
-            }
-        </style>
+        <link rel="stylesheet" type="text/css" href="./options.css"/>
     </head>
     <body>
         <div class="panel browser-style">
